+++ WALLET.md
@@ -442,9 +442,8 @@ If you have not yet created a keystore, one will be required to generate a relea
 
 ```sh
 cd android/app
-keytool -genkey -v -keystore celo-release-key.keystore -alias celo-key-alias -storepass celoFakeReleaseStorePass -keypass celoFakeReleaseKeyPass -keyalg RSA -keysize 2048 -validity 10000 -dname "CN=Android Debug,O=Android,C=US"
-export CELO_RELEASE_STORE_PASSWORD=celoFakeReleaseStorePass
-export CELO_RELEASE_KEY_PASSWORD=celoFakeReleaseKeyPass
+keytool -genkey -v -keystore mobilestack-release-key.keystore -alias mobilestack-key-alias -storepass fakeReleaseStorePass -keypass fakeReleaseStorePass -keyalg RSA -keysize 2048 -validity 10000 -dname "CN=Android Debug,O=Android,C=US"
+export RELEASE_STORE_PASSWORD=fakeReleaseStorePass
 ```
 
 ### Building an APK or Bundle
