+++ android/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>celo</name>
-	<comment>Project celo created by Buildship.</comment>
+	<name>mobilestack</name>
+	<comment>Project mobilestack created by Buildship.</comment>
 	<projects>
 	</projects>
 	<buildSpec>
