+++ android/app/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
 	<name>app</name>
-	<comment>Project app created by Buildship.</comment>
+	<comment>Project mobilestack created by Buildship.</comment>
 	<projects>
 	</projects>
 	<buildSpec>
