+++ android/app/mobilestack-release-key.keystore.enc
