+++ android/app/src/androidTest/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:tools="http://schemas.android.com/tools" package="org.celo.mobile" android:versionCode="1" android:versionName="1.0">
+  xmlns:tools="http://schemas.android.com/tools" package="xyz.mobilestack" android:versionCode="1" android:versionName="1.0">
   <uses-sdk android:minSdkVersion="21" android:targetSdkVersion="27" tools:overrideLibrary="com.wix.detox, android.support.test.uiautomator.v18" />
 </manifest>
