+++ android/app/src/androidTest/java/xyz/mobilestack/DetoxTest.java
@@ -1,4 +1,4 @@
-package org.celo.mobile;
+package xyz.mobilestack;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
@@ -25,7 +25,7 @@ public class DetoxTest {
     DetoxConfig detoxConfig = new DetoxConfig();
     detoxConfig.idlePolicyConfig.masterTimeoutSec = 90;
     detoxConfig.idlePolicyConfig.idleResourceTimeoutSec = 60;
-    detoxConfig.rnContextLoadTimeoutSec = (org.celo.mobile.BuildConfig.DEBUG ? 180 : 60);
+    detoxConfig.rnContextLoadTimeoutSec = (xyz.mobilestack.BuildConfig.DEBUG ? 180 : 60);
 
     Detox.runTests(mActivityRule, detoxConfig);
   }
