+++ android/app/src/main/AndroidManifest.xml
@@ -46,7 +46,7 @@
     <!-- Set custom default icon. This is used when no icon is set for incoming notification messages.
      See README(https://goo.gl/l4GJaQ) for more. -->
     <meta-data android:name="com.google.firebase.messaging.default_notification_icon" android:resource="@drawable/ic_notification_small" />
-    <meta-data tools:replace="android:resource" android:name="com.google.firebase.messaging.default_notification_color" android:resource="@color/celo_notification" />
+    <meta-data tools:replace="android:resource" android:name="com.google.firebase.messaging.default_notification_color" android:resource="@color/mobilestack_notification" />
 
     <meta-data android:name="CLEVERTAP_BACKGROUND_SYNC" android:value="1"/>
