+++ android/app/src/main/java/xyz/mobilestack/UserAgentClientFactory.java
@@ -1,4 +1,4 @@
-package org.celo.mobile;
+package xyz.mobilestack;
 
 import android.content.Context;
 import com.facebook.react.modules.network.OkHttpClientFactory;
