+++ android/app/src/main/java/xyz/mobilestack/UserAgentInterceptor.java
@@ -1,4 +1,4 @@
-package org.celo.mobile;
+package xyz.mobilestack;
 
 import android.os.Build;
 import java.io.IOException;
