+++ android/gradle.properties
@@ -18,8 +18,10 @@ org.gradle.jvmargs=-Xmx4096m -XX:+HeapDumpOnOutOfMemoryError
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 org.gradle.parallel=true
 
-CELO_RELEASE_STORE_FILE=celo-release-key.keystore
-CELO_RELEASE_KEY_ALIAS=celo-key-alias
+# Note that the key alias lacks Mobile Stack naming, since updating this would
+# require regenerating the keystore.
+RELEASE_STORE_FILE=release-key.keystore
+RELEASE_KEY_ALIAS=mobilestack-key-alias
 
 # Setting this manually based on version number until we have this deploying via Cloud Build
 # Example: v1.5.1 deployment number 1 = 1005001001 (1 005 001 001)
